Add Parser object with custom registry
Add the Parser struct, which just stores a reference to a function
registry. Give it `Parse` and `MustParse` methods, which parse paths
with the supplied registry. This allows the use of a registry with
custom extensions. Update the `Parse` function to delegate to a parser
and add a `MustParse` function to do the same but panic instead of
returning an error.

Change `parse.Parse` to take a registry as an argument, and teach the
Parser to pass one.

Add example tests demonstrating these use cases.

Update the golangci-lint config to allow functions that return
interfaces defined by the spec package and to ignore the double `R.` in
the `J. R. R. Tolkien" examples.