Properly lex shorthand name runes
Drop the JavaScript syntax enforcement adopted from theory/sqljson and
recognize the characters in a short hand identifier (`.foo`) as defined
by the JSONPath spec. This excludes support for escapes, so also
eliminate the code that handled escapes in identifiers. Update the tests
to reflect this change, especially the compliance test, which has been
modifying a test case to comply with the JavaScript syntax instead of
JSONPath. This gets us to 100% compatibility with the compliance test
suite.