Add support for TinyGo
Change the parsing of blank space so it doesn't create a 64 bit integer,
which TinyGo does not support. Add a basic app in `internal/wasm` and
`wasm` targets to the `Makefile` to compile it using both native go WASM
support and TinyGo. Add steps to the `.github/workflows/ci.yml` tests to
executes those targets, to help catch future changes that might break
WASM compatibility.