Fix linting issues
Add `.vscode/settings.json` to prevent VSCode from complaining about
syscall/js, use the correct schema for the GitHub workflow, and to
ignore the spelling of some words. Also use `GOOS=js GOARCH=wasm` in the
`Makefile` so that the compiler also is fine with
`.vscode/settings.json` while linting.