Rename ComparisonExpr, add Wildcard(), doc stable
Rename `ComparisonExpr` to `CompExpr` to match the naming of `CompVal`
and `CompOp`.

Replace the `Wildcard` global variable with the `Wildcard()`
constructor, which returns a singleton, but an be be changed in the
future if for some reason it needs not to be a singleton.

Document the stable types and constructors in the `spec` package, and
note that there may be some changes to other stuff in the future (though
probably not).

Link to the `Wildcard` global from the `WildcardSelector` docs and
document that `Registry.Get` is used internally by the parser.